src/requirements.txt
numpy
pillow
scipy
torch
torchvision
